% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.contact.delete.R
\name{uptimerobot.contact.delete}
\alias{uptimerobot.contact.delete}
\title{Delete an alert contact}
\usage{
uptimerobot.contact.delete(api.key, id)
}
\arguments{
\item{api.key}{string with a valid key for connecting to Uptimerobot API.}

\item{id}{numeric or integer with the ID of the contact to delete.}
}
\value{
The function returns \code{TRUE} in case success. An error is thrown otherwise.
}
\description{
\code{uptimerobot.contact.delete} removes an alert contanct, unlinking from all the registered monitors.
}
\examples{
\dontrun{
 # Let's assume the api.key is available into the environment variable KEY
 api.key <- Sys.getenv("KEY", "")

 # Create a new contact and get the ID
 if(uptimerobot.contact.delete(api.key, 12345678){
   message("Alert contact successfully deleted!")
 }
}
}
\author{
Gabriele Baldassarre
}

