% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.contacts.R
\name{uptimerobot.contacts}
\alias{uptimerobot.contacts}
\title{Get general informations about the registered contacts}
\usage{
uptimerobot.contacts(api.key, contacts = NA, limit = 50, offset = 0,
  fields = "id,value,friendlyname,type,status")
}
\arguments{
\item{api.key}{A valid key for connecting to UptimeRobors public API.}

\item{contacts}{vector or comma-delimited string with the IDs of the contacts to get.
If the argument is missing or NA, all the available monitors will be returned.}

\item{limit}{An integer value used for pagination. Defines the max number of records to return in each page. Default and max. is 50.}

\item{offset}{An integer value to set the index of the first monitor to get (used for pagination).}

\item{fields}{vector or comma-delimited string with the general informations to include in the output dataset.}
}
\description{
\code{uptimerobot.contacts} return a dataset with general informations
for a set of contacts used to be alert in case of uptimerobot events.
}
\details{
If a vector of contact IDs is not given, the function will return data for all the available contacts.

The API uses pagination and returns no more than 50 contacts on each page. Use \code{limit} and \code{offset} to set a different number of
monitors to get on each page and to move between pages. Leave default values to get all the data.
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{uptimerobot.monitors}}
}

