% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.monitor.new.R
\name{uptimerobot.monitor.new}
\alias{uptimerobot.monitor.new}
\title{Add a new monitor}
\usage{
uptimerobot.monitor.new(api.key, monitor.friendly.name, monitor.URL,
  monitor.type, monitor.subtype = NA, monitor.port = NA,
  monitor.interval = 5, monitor.keyword.type = NA,
  monitor.keyword.value = NA, monitor.HTTP.username = NA,
  monitor.HTTP.password = NA, monitor.alert.contacts = NA)
}
\arguments{
\item{api.key}{string with a valid key for connecting to Uptimerobot API.}

\item{monitor.friendly.name}{string the friendly (screen) name of the monitor.}

\item{monitor.URL}{string with the URL/IP of the monitor.}

\item{monitor.type}{string or integer with the type of the monitor. You can use both the friendly name (string)or the index (integer) here.}

\item{monitor.subtype}{string used only for "Port monitoring" to set which pre-defined port/service is monitored or if a custom port is monitored. You can use both the friendly name (string) or the index (integer) here.}

\item{monitor.port}{string used only for "Port monitoring" to set the port monitored.}

\item{monitor.interval}{integer with the interval for the monitoring check (in minutes)}

\item{monitor.keyword.value}{string with the value of the keyword (required for keyword monitoring).}

\item{monitor.HTTP.username}{string used for password-protected web pages (HTTP Basic Auth). Available for HTTP and keyword monitoring.}

\item{monitor.HTTP.password}{string used for password-protected web pages (HTTP Basic Auth). Available for HTTP and keyword monitoring.}

\item{monitor.alert.contacts}{character vector or data frame with the IDs to alert each with their threshold and recurrence values.}

\item{monitor.keyword,type}{required string in Keyword monitoring".}
}
\description{
\code{uptimerobot.monitor.new} create a new monitor with the given properties.
}
\details{
The function returns the ID of the newly created monitor in case success, an error is thrown otherwise.

The alert contacts are whom to be notified when the monitor goes up/down.

Multiple alert contact IDs can be sent in a character vector or in a data frame. If you pass alert contact IDs in a vector, each element must be formatted in the form \code{<id>_<threshold>_<recurrence>} (note the underscores).
If you prefer to format it as a data.frame, it must have these three columns: \code{id, threshold, recurrence}, numeric or integer. Order of the columns doesn't matter.

Please note that thresholds and recurrences can be omitted (default to zero) and, as they are only available in the Pro Plan, they are always 0 in the Free Plan.
}
\author{
Gabriele Baldassarre
}

