% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.monitor.reset.R
\name{uptimerobot.monitor.reset}
\alias{uptimerobot.monitor.reset}
\title{Delete a monitor}
\usage{
uptimerobot.monitor.reset(api.key, id)
}
\arguments{
\item{api.key}{string with a valid key for connecting to Uptimerobot API.}

\item{id}{numeric or integer with the ID of the monitor to delete.}
}
\value{
The function returns \code{TRUE} in case success. An error is thrown otherwise.
}
\description{
\code{uptimerobot.monitor.reset} remove all the statistics and logs associated to a monitor ID.
}
\author{
Gabriele Baldassarre
}

