% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobots.monitors.R
\name{uptimerobots.monitors}
\alias{uptimerobots.monitors}
\title{Extract monitors metadata and response times.}
\usage{
uptimerobots.monitors(api.key, monitors = NA, response.times = FALSE,
  fields = "id,friendlyname,url,type,port,interval,status")
}
\arguments{
\item{api.key}{A valid key for connecting to UptimeRobors public API.}

\item{monitors}{vector or comma-delimited string with the IDs of the monitors to get. If the argument is missing, all monitors will be get.}

\item{response.times}{Set to TRUE to get the response times with the metadata. FALSE to get  metadata values only.}

\item{fields}{vector or comma-delimited string with the columns of metadata to get.
\code{n}.}
}
\description{
\code{uptimerobots.monitors} get the available response times and metadata of the given \code{monitors} vectors.
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{uptimerobots.monitors}}
}

