% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.contact.delete.R
\name{uptimerobot.contact.delete}
\alias{uptimerobot.contact.delete}
\title{Delete an alert contact}
\usage{
uptimerobot.contact.delete(api.key, id)
}
\arguments{
\item{api.key}{string with a valid key for connecting to Uptimerobot API.}

\item{id}{numeric or integer with the ID of the contact to delete.}
}
\description{
\code{uptimerobot.contact.delete} remove an alert contanct, unlinking from all the registered monitors.

The function returns \code{TRUE} in case success. An error is thrown otherwise.
}
\author{
Gabriele Baldassarre
}

