% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.contact.new.R
\name{uptimerobot.contact.new}
\alias{uptimerobot.contact.new}
\title{Add a new alert contact}
\usage{
uptimerobot.contact.new(api.key, type, value, friendly.name)
}
\arguments{
\item{api.key}{string with a valid key for connecting to Uptimerobot API.}

\item{type}{string or integer with the type of the contact. You can use both the friendly name (string) or the index (integer) here.}

\item{value}{string with the value of the contact (ie. the email address).}

\item{friendly.name}{string the friendly (screen) name of the contact.}
}
\description{
\code{uptimerobot.contact.new} creates a new alert contact with the given properties.
}
\details{
The alert contacts are whom to be notified when the monitor goes up/down.

The function returns the ID of the newly created contact in case success. An error is thrown otherwise.
}
\author{
Gabriele Baldassarre
}

