% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.monitor.delete.R
\name{uptimerobot.monitor.delete}
\alias{uptimerobot.monitor.delete}
\title{Delete a monitor}
\usage{
uptimerobot.monitor.delete(api.key, id)
}
\arguments{
\item{api.key}{string with a valid key for connecting to Uptimerobot API.}

\item{id}{numeric or integer with the ID of the monitor to delete.}
}
\value{
The function returns \code{TRUE} in case success. An error is thrown otherwise.
}
\description{
\code{uptimerobot.monitor.delete} remove a monitor and all existing statistics of it.
}
\author{
Gabriele Baldassarre
}

