% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uptimerobot.monitors.R
\name{uptimerobot.monitors}
\alias{uptimerobot.monitors}
\title{Get general informations for one or more monitors}
\usage{
uptimerobot.monitors(api.key, monitors = NA, types = NA, statuses = NA,
  search = NA, summary = list(), limit = 50, offset = 0,
  fields = uptimerobot.fields("monitor")$typical)
}
\arguments{
\item{api.key}{A valid key for connecting to UptimeRobors public API.}

\item{monitors}{vector or comma-delimited string with the IDs of the monitors to get.
If the argument is missing or NA, all the available monitors will be returned.}

\item{types}{vector or comma-delimited string of monitor types. If not \code{NA}, the function will return all monitors types (HTTP, keyword, ping..) in an account. Else, it is possible to define any number of monitor types.}

\item{search}{An optional keyword of to search within monitor.url and monitor.friendly.name to get filtered results.}

\item{summary}{list of logical values to flag summary indicators to add to the output dataset.}

\item{limit}{An integer value used for pagination. Defines the max number of records to return in each page. Default and max. is 50.}

\item{offset}{An integer value to set the index of the first monitor to get (used for pagination).}

\item{fields}{vector or comma-delimited string with the general informations to include in the output dataset.
You may use the helper function \code{\link{uptimerobot.fields}} if you don't want to manually compile the list of fields.}

\item{statutes}{vector or comma-delimited string of monitor statutes. If not \code{NA}, the function will return  all monitors statuses (up, down, paused) in an account. Else, it is possible to define any number of monitor statuses.}
}
\description{
\code{uptimerobots.monitors.responses} return a dataset with general informations
for a set of monitors.
}
\details{
If a vector of monitor is not given, the function will return data for all the available monitors.

\code{summary} parameter expect a lists of three named logic values that set which columns of additional statistics for each monitor must  be added to output dataset for
each available monitor:

\enumerate{
  \item \code{response.times} set to \code{TRUE} to add the number of pings with response times available for the monitor. These values can be queried using \code{\link{uptimerobot.monitor.responses}} function.
  \item \code{log.records} set to \code{TRUE} to add the number of log entries recorded for the monitor. These records can be queried using \code{\link{uptimerobot.monitor.logs}} function.
  \item \code{alert.contacts} set to \code{TRUE} to add number of alert contacts binded to the monitor. Detailed informations about these contacts can be queried using \code{\link{uptimerobot.monitor.contacts}} function.
}

You may just add the elements you want to include into the list, as they default to \code{FALSE} if missing. Set an empty list to exclude all the summary statistics from the output.

The API uses pagination and returns no more than 50 monitors on each page. Use \code{limit} and \code{offset} to set a different number of
monitors to get on each page and to move between pages. Leave default values to get all the data.
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{uptimerobot.monitor.responses}}, \code{\link{uptimerobot.monitor.logs}}, \code{\link{uptimerobot.monitor.contacts}}
}

